Move EnvironmentVariablesTrait
